requirements.txt
newsapi-python==0.2.7
google-genai==0.2.2
python-dotenv==1.0.1
Pillow==11.0.0
